+++ inst/doc/POMDP.R
@@ -6,18 +6,6 @@ library("pomdp")
 str(args(POMDP))
 
 ## ---- eval = FALSE-------------------------------------------------------
-#  discount = 0.9
-
-## ---- eval = FALSE-------------------------------------------------------
-#  states = c("state1" , "state2" , "state3")
-
-## ---- eval=FALSE---------------------------------------------------------
-#  actions = c("action1" , "action2")
-
-## ---- eval=FALSE---------------------------------------------------------
-#  observations = c("obs1" , "obs2")
-
-## ---- eval = FALSE-------------------------------------------------------
 #  start = c(0.5 , 0.3 , 0.2)
 
 ## ---- eval = FALSE-------------------------------------------------------
@@ -35,36 +23,11 @@ str(args(POMDP))
 #  start = c("-" , "state2")
 
 ## ---- eval = FALSE-------------------------------------------------------
-#  transition_prob = data.frame(
-#    "action" = c(
-#            "action1" , "action1" , "action1",
-#            "action1" , "action1" , "action1",
-#            "action1" , "action1" , "action1",
-#            "action2" , "action2" , "action2",
-#            "action2" , "action2" , "action2",
-#            "action2" , "action2" , "action2"),
-#    "start-state" = c(
-#            "state1" , "state1" , "state1",
-#            "state2" , "state2" , "state2",
-#            "state3" , "state3" , "state3",
-#            "state1" , "state1" , "state1",
-#            "state2" , "state2" , "state2",
-#            "state3" , "state3" , "state3"),
-#     "end-state" = c(
-#            "state1" , "state2" , "state3",
-#            "state1" , "state2" , "state3",
-#            "state1" , "state2" , "state3",
-#            "state1" , "state2" , "state3",
-#            "state1" , "state2" , "state3",
-#            "state1" , "state2" , "state3"),
-#     "probability" = c(
-#            0.1  0.4, 0.5,
-#            0,   0.7, 0.3,
-#            0.4, 0.4, 0.2,
-#            0,   0.6, 0.4,
-#            0.1, 0.9, 0,
-#            0.7, 0.3, 0)
-#      )
+#  transition_prob = rbind(
+#    T_("action1", "state1", "state2", 0.1),
+#    T_("action2", "state1", "state3", 0.9),
+#    T_("*"      , "state2", "*",      1)
+#  )
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  transition_prob = list(
@@ -85,11 +48,14 @@ str(args(POMDP))
 #      "action2" = "uniform")
 
 ## ---- eval = FALSE-------------------------------------------------------
-#  observation_prob = data.frame(
-#    "action"      = c(     "*",      "*",      "*",      "*",      "*",      "*"),
-#    "end-state"   = c("state1", "state1", "state2", "state2", "state3", "state3"),
-#    "observation" = c(  "obs1",   "obs2",   "obs1",   "obs2",   "obs1",   "obs2"),
-#    "probability" = c(     0.1,      0.9,      0.3,      0.7,      0.4,      0.6))
+#  observation_prob = rbind(
+#    O_("*", "state1", "obs1", 0.1),
+#    O_("*", "state1", "obs2", 0.9),
+#    O_("*", "state2", "obs1", 0.3),
+#    O_("*", "state2", "obs2", 0.7),
+#    O_("*", "state3", "obs1", 0.5),
+#    O_("*", "state3", "obs2", 0.6)
+#  )
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  observation_prob = list(
@@ -108,12 +74,12 @@ str(args(POMDP))
 #                        0.4, 0.6), nrow = 3, byrow = TRUE))
 
 ## ---- eval = FALSE-------------------------------------------------------
-#  reward = data.frame(
-#    "action" = c("action1", "action1", "action1", "action2", "action2", "action2"),
-#    "start-state" = c("*", "*", "*", "*", "*", "*"),
-#    "end-state" = c("state1", "state2", "state3", "state1", "state2", "state3"),
-#    "observation" = c("*", "*", "*", "*", "*", "*") ,
-#    "reward" = c(10000, 2000, 50, 150, 2500, 100))
+#  reward = rbind(
+#    R_("action1", "*", "state1", "*", 10000),
+#    R_("action1", "*", "state2", "*", 2000),
+#    R_("action2", "*", "state1", "*", 50),
+#    R_("action2", "*", "state2", "*", 100)
+#  )
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  reward = list(
@@ -125,13 +91,6 @@ str(args(POMDP))
 #       "state1" = matrix(c(3, 2, 4, 7, 4, 8) , nrow = 3 , byrow = TRUE),
 #       "state2" = matrix(c(0, 9, 8, 2, 5, 4) , nrow = 3 , byrow = TRUE),
 #       "state3" = matrix(c(4, 3, 4, 4, 5, 6) , nrow = 3 , byrow = TRUE)))
-
-## ---- eval = FALSE-------------------------------------------------------
-#  values = "cost"
-#  values = "reward"
-
-## ---- eval = FALSE-------------------------------------------------------
-#  name = "Test Problem"
 
 ## ------------------------------------------------------------------------
 str(args(solve_POMDP))
@@ -160,14 +119,15 @@ TigerProblem <- POMDP(
     "open-left" = "uniform",
     "open-right" = "uniform"),
     
-  reward = data.frame(
-    "action" = c("listen", "open-left", "open-left", "open-right", "open-right"),
-    "start-state" = c("*", "tiger-left", "tiger-right", "tiger-left", "tiger-right"),
-    "end-state" = c("*", "*", "*", "*", "*"),
-    "observation" = c("*", "*", "*", "*", "*"),
-    "reward" = c(-1, -100, 10, 10, -100))
+  reward = rbind(
+    R_("listen",     "*",           "*", "*", -1  ),
+    R_("open-left",  "tiger-left",  "*", "*", -100),
+    R_("open-left",  "tiger-right", "*", "*", 10  ),
+    R_("open-right", "tiger-left",  "*", "*", 10  ),
+    R_("open-right", "tiger-right", "*", "*", -100)
   )
-  
+)
+
 TigerProblem
 
 ## ------------------------------------------------------------------------
@@ -184,7 +144,9 @@ plot(tiger_solved)
 alpha <- solution(tiger_solved)$alpha
 alpha
 
-plot(NA, xlim = c(0, 1), ylim = c(0, 10), xlab = "Belief space", ylab = "Value function")
-for(i in 1:nrow(alpha)) abline(a = alpha[i,1], b= alpha[i,2], col = i)
-legend("topright", legend = 1:nrow(alpha), col = 1:nrow(alpha), lwd=1)
+plot(NA, xlim = c(0, 1), ylim = c(0, 20), xlab = "Belief space (for tiger is left)", 
+  ylab = "Value function")
+for(i in 1:nrow(alpha)) abline(a = alpha[i,2], b = alpha[i,1]- alpha[i,2], col = i, xpd = FALSE)
+legend("topright", legend = 
+    paste0(1:nrow(alpha),": ", solution(tiger_solved)$pg[,"action"]), col = 1:nrow(alpha), lwd=1)