+++ NEWS.md
@@ -1,7 +1,21 @@
+# pomdp 0.9.2 (12/06/2019)
+
+## Changes
+* solve_POMDP can now solve POMDP files.
+* added helper functions O, R and T.
+* improved plot.
+* Added reward function.
+* values argument is now called max.
+* Fixed class structure. The central class is not POMDP with elements model and solution.
+
+## Bugfix
+* fixed warning for start = "uniform".
+* fixed warning in C code for gcc10.
+
 # pomdp 0.9.1-1 (05/14/2019)
 
 ## Bugfix
-* fixed warning in mdp.c for gcc9
+* fixed warning in mdp.c for gcc9.
 
 # pomdp 0.9.1 (01/02/2019)